ProjectX/requirements.txt
pyTelegramBotAPI
google-genai
numpy
